moved query_profile model tests to new file
