+++ tests/unit/test_query.py
@@ -1501,79 +1501,3 @@ def _make_datastore_api(*results):
 
     return mock.Mock(run_query=run_query, spec=["run_query"])
 
-
-def test_explain_metrics__from_pb():
-    from google.cloud.datastore.query_profile import (
-        ExplainMetrics,
-        _ExplainAnalyzeMetrics,
-        QueryExplainError,
-        PlanSummary,
-    )
-    from google.cloud.datastore_v1.types import query_profile as query_profile_pb2
-    from google.protobuf import struct_pb2, duration_pb2
-
-    # test without execution_stats field
-    expected_metrics = query_profile_pb2.ExplainMetrics(
-        plan_summary=query_profile_pb2.PlanSummary(
-            indexes_used=struct_pb2.ListValue(values=[])
-        )
-    )
-    metrics = ExplainMetrics._from_pb(expected_metrics)
-    assert isinstance(metrics, ExplainMetrics)
-    assert isinstance(metrics.plan_summary, PlanSummary)
-    assert metrics.plan_summary.indexes_used == []
-    with pytest.raises(QueryExplainError) as exc:
-        metrics.execution_stats
-    assert "execution_stats not available when explain_options.analyze=False" in str(
-        exc.value
-    )
-    # test with execution_stats field
-    expected_metrics.execution_stats = query_profile_pb2.ExecutionStats(
-        results_returned=1,
-        execution_duration=duration_pb2.Duration(seconds=2),
-        read_operations=3,
-        debug_stats=struct_pb2.Struct(
-            fields={"foo": struct_pb2.Value(string_value="bar")}
-        ),
-    )
-    metrics = ExplainMetrics._from_pb(expected_metrics)
-    assert isinstance(metrics, ExplainMetrics)
-    assert isinstance(metrics, _ExplainAnalyzeMetrics)
-    assert metrics.execution_stats.results_returned == 1
-    assert metrics.execution_stats.execution_duration.total_seconds() == 2
-    assert metrics.execution_stats.read_operations == 3
-    assert metrics.execution_stats.debug_stats == {"foo": "bar"}
-
-
-def test_explain_metrics_execution_stats():
-    """
-    Standard ExplainMetrics class should raise exception when execution_stats is accessed.
-    _ExplainAnalyzeMetrics should include the field
-    """
-    from google.cloud.datastore.query_profile import (
-        ExplainMetrics,
-        QueryExplainError,
-        _ExplainAnalyzeMetrics,
-    )
-
-    metrics = ExplainMetrics(plan_summary=object())
-    with pytest.raises(QueryExplainError) as exc:
-        metrics.execution_stats
-    assert "execution_stats not available when explain_options.analyze=False" in str(
-        exc.value
-    )
-    expected_stats = object()
-    metrics = _ExplainAnalyzeMetrics(
-        plan_summary=object(), _execution_stats=expected_stats
-    )
-    assert metrics.execution_stats is expected_stats
-
-
-def test_explain_options__to_dict():
-    """
-    Should be able to create a dict representation of ExplainOptions
-    """
-    from google.cloud.datastore.query_profile import ExplainOptions
-
-    assert ExplainOptions(analyze=True)._to_dict() == {"analyze": True}
-    assert ExplainOptions(analyze=False)._to_dict() == {"analyze": False}
